requirements.txt
asgiref==3.6.0
attrs==22.2.0
autopep8==2.0.1
certifi==2022.12.7
charset-normalizer==2.0.12
click==8.1.3
coverage==7.1.0
flake8==6.0.0
Flask==2.2.2
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.2
mccabe==0.7.0
numpy==1.24.1
oauthlib==3.2.2
packaging==23.0
pluggy==1.0.0
py==1.11.0
pycodestyle==2.10.0
pyflakes==3.0.1
pytest==7.2.1
python-decouple==3.7
python-dotenv==0.21.1
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
sqlparse==0.4.3
threadpoolctl==3.1.0
tomli==2.0.1
urllib3==1.26.14
Werkzeug==2.2.2
